requirements.txt
requests
beautifulsoup4
unittest
textblob
pandas
scikit-learn
google-api-python-client
pyyaml